refactor(campaign): decompose dayAdvancement.ts into 4 modules
dayAdvancement.ts had grown to 605 LOC, exceeding the critical
threshold for a single processor-chain file. Extract the three
cohesive day-cycle phases plus the shared report types into
co-located sibling modules, leaving dayAdvancement.ts as a thin
pipeline coordinator (605 -> 309 LOC).

New sibling modules:
- dayReportTypes.ts       (105 LOC) report/event interfaces + constants
- healingProcessing.ts    (147 LOC) processHealing personnel phase
- contractProcessing.ts   (64 LOC)  processContracts missions phase
- dailyCostsProcessing.ts (119 LOC) processDailyCosts finance phase

Behavior-preserving: the phase logic is a pure cut/paste with zero
functional change. dayAdvancement.ts re-exports every previously
exported symbol (DEFAULT_DAILY_SALARY, DEFAULT_DAILY_MAINTENANCE,
HealedPersonEvent, ExpiredContractEvent, DailyCostBreakdown,
TurnoverDepartureEvent, DayReport, processHealing, processContracts,
processDailyCosts) so all importers and tests keep working against
@/lib/campaign/dayAdvancement unchanged. Processor ordering and phase
sequencing are untouched.

Verification: tsc --noEmit clean, oxlint 0 warnings/0 errors,
275 campaign tests pass (54 dayAdvancement/dayPipeline + 221 processor).
